indexer

+++ sites/logout.php
@@ -1,5 +1,5 @@
 <?php
-	checkIndex();
+	if(!isset($fromIndex)): header("Location:".getURL()); exit; endif;
 
 	$mysql->execute("DELETE FROM `sessions` WHERE `sid` = ?", "s", session_id());
 	header("Location:".getURL());
